requirements.txt
telethon==1.27.0
openai==0.27.0
python-dotenv==0.19.2